+++ doc/version_history.rst
@@ -6,6 +6,20 @@
 Version History
 ###############
 
+v2.14.0
+-------
+
+* Updated the ScriptQueue behavior to move scripts to the "done" list and start running the next script as soon as the current script state is done, instead of waiting for the process to be done.
+
+v2.13.2
+-------
+
+* In block_model, fix issue with remove_done_blocks.
+
+* In block_info, fix block_type when block is a test case.
+
+* In block_info, fix done() method.
+
 v2.13.1
 -------